test(LottoGeneratorTest.java) : 중복 테스트 병합
